requirements_fixed.txt
# MedAI Radiologia - Dependências Corrigidas
# Core ML/AI
tensorflow>=2.10.0
numpy>=1.21.0
scikit-learn>=1.1.0
opencv-python>=4.6.0
Pillow>=9.0.0

# Medical Imaging
pydicom>=2.3.0
nibabel>=4.0.0
SimpleITK>=2.2.0

# Web Frameworks
flask>=2.2.0
flask-cors>=3.0.10
fastapi>=0.85.0
uvicorn[standard]>=0.18.0
pydantic>=1.10.0

# GUI (Optional)
PyQt5>=5.15.0

# Data Processing
pandas>=1.5.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
requests>=2.28.0
python-multipart>=0.0.5
aiofiles>=0.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Development
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0
